requirements.txt
numpy

imageio
scipy
argparse
tensorboardX
blessings
progressbar2
path.py